Implementation of matrix mutlipication in serial

+++ OpenMP/integration/include/trapezoidal_integration.h
@@ -25,7 +25,7 @@ private:
 	Number upper_limit;
 	Number lower_limit;
 	Number sum;
-	int intervals;
+	long int intervals;
 	Number step_size;
 
 public:
